requirements.txt
google-genai
streamlit
pandas
python-dotenv
